pyproject.toml
[tool.poetry]
name = "tamatem-tasks"
version = "0.1.0"
description = ""
authors = ["YahyaOmari <yahya.omari@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
matplotlib = "^3.4.2"
numpy = "^1.21.1"
pandas = "^1.3.0"
psycopg2 = "^2.9.1"
python-dotenv = "^0.18.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
